requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1